+++ R/calculate_distance.R
@@ -22,15 +22,8 @@
 #' calculate_distance(
 #'   raster = sr,
 #'   access = ac,
-#'   plot = TRUE
 #' )
 #'
-#' calculate_distance(
-#'   raster = sr,
-#'   access = ac,
-#'   plot = TRUE,
-#'   filename = tempfile(fileext = ".tif")
-#' )
 #' @author Tristan R.H. Goodbody
 #'
 #' @export