+++ R/calculate_lhsOpt.R
@@ -66,7 +66,7 @@ calculate_lhsOpt <- function(popLHS,
   }
 
   if (any(!names(popLHS) %in% c("values", "pcaLoad", "matQ", "matCov"))) {
-    stop(glue::glue("'popLHS' must be the output from the 'sgsR::calculate_pop()' function"))
+    stop(paste0("'popLHS' must be the output from the 'sgsR::calculate_pop()' function"))
   }
 
   if (!is.logical(PCA)) {
@@ -297,7 +297,7 @@ plot_LHCOptim <- function(dfFinal,
     dplyr::select(df.x) %>%
     dplyr::pull()
 
-  message(glue::glue("Your optimum estimated sample size based on KL divergence is: {num}"))
+  message(paste0("Your optimum estimated sample size based on KL divergence is: ", num))
 
   x2 <- c(num, num)
   y2 <- c(0, 1)