+++ R/extract_metrics.R
@@ -85,7 +85,7 @@ extract_metrics <- function(mraster,
       }
       
       if (file.exists(filename) & isFALSE(overwrite)) {
-        stop(glue::glue("{filename} already exists and overwrite = FALSE"))
+        stop(paste0("'",filename, "' already exists and overwrite = FALSE"))
       }
       
       sf::st_write(samples, filename, delete_layer = overwrite)