+++ R/extract_strata.R
@@ -122,7 +122,7 @@ extract_strata <- function(sraster,
       }
       
       if (file.exists(filename) & isFALSE(overwrite)) {
-        stop(glue::glue("{filename} already exists and overwrite = FALSE"))
+        stop(paste0("'",filename, "' already exists and overwrite = FALSE"))
       }
       
       sf::st_write(samples, filename, delete_layer = overwrite)