+++ R/sample_systematic.R
@@ -63,14 +63,6 @@
 #' sample_systematic(
 #'   raster = sr,
 #'   cellsize = 1000,
-#'   access = ac,
-#'   buff_inner = 50,
-#'   buff_outer = 200
-#' )
-#'
-#' sample_systematic(
-#'   raster = sr,
-#'   cellsize = 1000,
 #'   square = FALSE,
 #'   location = "random"
 #' )
@@ -290,7 +282,7 @@ sample_systematic <- function(raster,
     }
     
     if (file.exists(filename) & isFALSE(overwrite)) {
-      stop(glue::glue("{filename} already exists and overwrite = FALSE"))
+      stop(paste0("'",filename, "' already exists and overwrite = FALSE"))
     }
     
     sf::st_write(samples, filename, delete_layer = overwrite)