+++ R/strat_breaks.R
@@ -202,8 +202,8 @@ strat_breaks <- function(mraster,
     p <- ggplot2::ggplot(data, ggplot2::aes(val)) +
       ggplot2::geom_histogram() +
       ggplot2::geom_vline(xintercept = breaks, linetype = "dashed") +
-      ggplot2::ggtitle(glue::glue("{nm} histogram with defined breaks")) +
-      ggplot2::xlab(glue::glue("{nm}"))
+      ggplot2::ggtitle(paste0(nm, " histogram with defined breaks")) +
+      ggplot2::xlab(paste0(nm))
 
     if (!is.null(mraster2)) {
       data2 <- terra::as.data.frame(rastermetric2)