+++ R/zzz.R
@@ -1,4 +1,3 @@
-#' @import Rcpp
 .onAttach <- function(libname, pkgname) {
   # Runs when attached to search() path such as by library() or require()
   if (!interactive()) {