+++ NEWS.md
@@ -1,3 +1,9 @@
+# sgsR 0.1.2
+
+* Removed dependencies for `stringr`, `tidyselect`, `rlang`, `RColorBrewer`, `magrittr`, `glue`, `Rcpp`
+
+* Improved documentation examples to have read speeds in line with <5 seconds for CRAN submission
+
 # sgsR 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.